pyproject.toml
[project]
name = "mkdocs-material2014"
version = "0.4.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs>=1.6.1",
    "pymdown-extensions>=10.16.1",
    "mkdocs-macros-plugin==1.3.7",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
extras = [
    "mkdocs-literate-nav==0.6.2",
    "mkdocs-ezglossary-plugin>=2.0.0",
    "mkdocs-mermaid2-plugin>=1.2.1",
]
openapi = [
    "essentials-openapi>=1.2.1",
    "rich>=14.1.0",
    "httpx>=0.28.1",
    "pyyaml"
]

[build-system]
requires = ["setuptools >= 80.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[project.entry-points."mkdocs.themes"]
material2014 = "material2014.templates"
